change tests, change get_next_weekday to include today
