+++ tablefor2/helpers.py
@@ -61,8 +61,7 @@ def calculate_recurring_values(recurring):
 
 # find the next weekday after a given day, plus 7
 # https://stackoverflow.com/questions/6558535/find-the-date-for-the-first-monday-after-a-given-a-date
-def get_next_weekday(day, time):
-    today = datetime.datetime.utcnow().date()
+def get_next_weekday(today, day, time):
     days_ahead = int(day) - today.weekday()
     next_weekday = today + datetime.timedelta(days_ahead+7)  # 2017-07-19 12:00:00+00:00
