+++ tablefor2/test/test_match_for_lunch.py
@@ -1,0 +1,8 @@
+# from django.test import TestCase
+
+# from tablefor2.helpers import calculate_utc, determine_ampm, get_next_weekday
+# from tablefor2.models import *
+# from tablefor2.management.commands.match_users import Command
+
+# import datetime
+# import pytz
