+++ tablefor2/test/test_match_users.py
@@ -1,7 +1,5 @@
-from django.core.management import call_command
 from django.test import TestCase
 
-from tablefor2.helpers import calculate_utc, determine_ampm, get_next_weekday
 from tablefor2.models import *
 from tablefor2.management.commands.match_users import Command
 
@@ -767,153 +765,3 @@ class MatchTestCase(TestCase):
 
         self.assertEqual(Command.check_accept_matches(Command(), t, a), True)
         self.assertEqual(Command.check_accept_matches(Command(), poop, a), False)
-
-
-class RecurringAvailabilityTest(TestCase):
-    def setup_profiles(self):
-        # tiffany, Success, SF, No, once a month
-        Profile.objects.create(
-            first_name='tiffany',
-            last_name='qi',
-            preferred_first_name='tiffany',
-            email='tiffany@TEST-mixpanel.com',
-            department='Success',
-            location='San Francisco',
-            timezone='PST',
-            google_hangout='No',
-            frequency='Once a month',
-            accept_matches='Yes',
-            date_entered_mixpanel=datetime.datetime(2016, 10, 31),
-            distinct_id='tiffany'
-        )
-
-    def setup_recurring(self):
-        self.setup_profiles()
-        RecurringAvailability.objects.create(
-            profile=Profile.objects.get(first_name='tiffany'),
-            day="0",  # mon
-            time="1:00PM"
-        )
-        RecurringAvailability.objects.create(
-            profile=Profile.objects.get(first_name='tiffany'),
-            day="2",  # wed
-            time="12:00PM"
-        )
-        RecurringAvailability.objects.create(
-            profile=Profile.objects.get(first_name='tiffany'),
-            day="4",  # fri
-            time="10:00AM"
-        )
-
-    # you'll have to change this
-    def test_create_availabilities(self):
-        self.setup_recurring()
-        t = Profile.objects.get(first_name='tiffany')
-        av1 = Availability.objects.create(
-            profile=t,
-            time_available=datetime.datetime(2017, 8, 7, 13, 0, tzinfo=pytz.UTC),
-            time_available_utc=datetime.datetime(2017, 8, 7, 20, 0, tzinfo=pytz.UTC),
-        )
-        av2 = Availability.objects.create(
-            profile=t,
-            time_available=datetime.datetime(2017, 8, 9, 12, 0, tzinfo=pytz.UTC),
-            time_available_utc=datetime.datetime(2017, 8, 9, 19, 0, tzinfo=pytz.UTC),
-        )
-        av3 = Availability.objects.create(
-            profile=t,
-            time_available=datetime.datetime(2017, 8, 11, 10, 0, tzinfo=pytz.UTC),
-            time_available_utc=datetime.datetime(2017, 8, 11, 17, 0, tzinfo=pytz.UTC),
-        )
-        self.assertEqual([av1, av2, av3], Command.create_availabilities(Command()))
-
-    # you'll have to change this
-    def test_delete_availabilities(self):
-        self.setup_recurring()
-        t = Profile.objects.get(first_name='tiffany')
-        RecurringAvailability.objects.get(profile=t, day="0", time="1:00PM").delete()
-        av1 = Availability(
-            profile=t,
-            time_available=datetime.datetime(2017, 7, 24, 13, 0, tzinfo=pytz.UTC),
-            time_available_utc=datetime.datetime(2017, 7, 24, 20, 0, tzinfo=pytz.UTC),
-        )
-        av2 = Availability(
-            profile=t,
-            time_available=datetime.datetime(2017, 7, 31, 13, 0, tzinfo=pytz.UTC),
-            time_available_utc=datetime.datetime(2017, 7, 31, 20, 0, tzinfo=pytz.UTC),
-        )
-        # self.assertEqual([av1, av2], Command.delete_availabilities(Command()))  # why not equal??
-
-
-class HelpersTest(TestCase):
-    def setup_profiles(self):
-        # tiffany, Success, SF, No, once a month
-        Profile.objects.create(
-            first_name='tiffany',
-            last_name='qi',
-            preferred_first_name='tiffany',
-            email='tiffany@TEST-mixpanel.com',
-            department='Success',
-            location='San Francisco',
-            timezone='PST',
-            google_hangout='No',
-            frequency='Once a month',
-            date_entered_mixpanel=datetime.datetime(2016, 10, 31),
-            distinct_id='tiffany'
-        )
-        # Tim, Engineering, New York, Yes, once a month
-        Profile.objects.create(
-            first_name='tim',
-            last_name='trefen',
-            preferred_first_name='tim',
-            email='tim@TEST-mixpanel.com',
-            department='Engineering',
-            location='New York',
-            google_hangout='Yes',
-            timezone='EST',
-            frequency='Once a month',
-            date_entered_mixpanel=datetime.datetime(2013, 06, 01),
-            distinct_id='tim'
-        )
-        # Karima, Success, Other, Yes, once a month
-        Profile.objects.create(
-            first_name='karima',
-            last_name='el moujahid',
-            preferred_first_name='karima',
-            email='karima@TEST-mixpanel.com',
-            department='Success',
-            location='Other',
-            google_hangout='Yes',
-            timezone='CEST',
-            frequency='Once a month',
-            date_entered_mixpanel=datetime.datetime(2016, 06, 01),
-            distinct_id='karima'
-        )
-
-    def test_calculate_utc(self):
-        self.setup_profiles()
-        t = Profile.objects.get(first_name='tiffany')
-        tim = Profile.objects.get(first_name='tim')
-        k = Profile.objects.get(first_name='karima')
-
-        # not daylight savings
-        self.assertEqual(calculate_utc(t, datetime.datetime(2016, 1, 5, 12, 0)), datetime.datetime(2016, 1, 5, 20, 0, tzinfo=pytz.UTC))
-        self.assertEqual(calculate_utc(tim, datetime.datetime(2016, 1, 5, 15, 0)), datetime.datetime(2016, 1, 5, 20, 0, tzinfo=pytz.UTC))
-        self.assertEqual(calculate_utc(k, datetime.datetime(2016, 1, 5, 21, 0)), datetime.datetime(2016, 1, 5, 20, 0, tzinfo=pytz.UTC))
-
-        # daylight savings
-        self.assertEqual(calculate_utc(t, datetime.datetime(2017, 5, 21, 11, 30)), datetime.datetime(2017, 5, 21, 18, 30, tzinfo=pytz.UTC))
-        self.assertEqual(calculate_utc(tim, datetime.datetime(2017, 5, 21, 14, 30)), datetime.datetime(2017, 5, 21, 18, 30, tzinfo=pytz.UTC))
-        self.assertEqual(calculate_utc(k, datetime.datetime(2017, 5, 21, 20, 30)), datetime.datetime(2017, 5, 21, 18, 30, tzinfo=pytz.UTC))
-
-    def test_determine_ampm(self):
-        self.assertEqual(determine_ampm('12:00PM'), '12:00')
-        self.assertEqual(determine_ampm('5:00PM'), '17:00')
-        self.assertEqual(determine_ampm('1:00PM'), '13:00')
-        self.assertEqual(determine_ampm('10:00AM'), '10:00')
-        self.assertEqual(determine_ampm('9:30AM'), '9:30')
-
-    # you'll have to change this
-    def test_get_next_weekday(self):
-        self.assertEqual(get_next_weekday('0', '12:00'), datetime.datetime(2017, 8, 7, 12, 0))
-        self.assertEqual(get_next_weekday('1', '1:00'), datetime.datetime(2017, 8, 8, 1, 0))
-        self.assertEqual(get_next_weekday('4', '15:30'), datetime.datetime(2017, 8, 11, 15, 30))
